Multiplayer: The Huge PvP Ban
Ignore-this: 9f267fdbd53cae93cdd3d957225b9200
Whew, this was a lot of work. Hopefully this prevents all possible
interaction between players, other than indirect interaction like
dropping items for another player to pick up. There are probably
bugs in this, though, because it's so big. (Another exception; a
few player/player interactions, like casting healing at each other,
are special-cased to be legal.) This is the major change needed for
multiplayer to actually work (another important change that will be
required, but is not included in this patch, will be to improve
monster AI to handle the new rules). It also bans other nondriving
player interactions, such as a monster hitting both players in one
attack, or a monster aiming at one player and hitting the other,
and adds a huge number of sanity checks to catch multiplayer getting
out of synch.

darcs-hash:20110720200815-99289-96f7f5232d6d3d588f2e01c1d603ca14b1daaed9