Add roomcolors option.
If the user enables the roomcolors option, some rooms (mainly in the
main dungeon) will have colored walls.  The color varies from room to
room in a deterministic but apparently arbitrary manner.  One out of
every twenty colored rooms is a "rainbow room".

This is purely cosmetic and has no impact on gameplay.  In fact,
the option can be toggled in-game, though you may need to force a
redraw to see the change immediately.